+++ hasklepias-core/src/Tests/Cohort/Attrition.hs
@@ -11,25 +11,22 @@ import           Test.Tasty
 import           Test.Tasty.HUnit
 
 testAttr1 :: AttritionInfo
-testAttr1 = MkAttritionInfo 1 2 $ fromList
-  [ MkAttritionLevel SubjectHasNoIndex         0
-  , MkAttritionLevel (ExcludedBy (1, "feat2")) 1
-  , MkAttritionLevel Included                  1
-  ]
+testAttr1 = makeTestAttritionInfo 1 2 
+  [ (SubjectHasNoIndex, 0)
+  , (ExcludedBy (1, "feat2"), 1)
+  , (Included, 1)]
 
 testAttr2 :: AttritionInfo
-testAttr2 = MkAttritionInfo 1 5 $ fromList
-  [ MkAttritionLevel SubjectHasNoIndex         0
-  , MkAttritionLevel (ExcludedBy (1, "feat2")) 3
-  , MkAttritionLevel Included                  2
-  ]
-
+testAttr2 = makeTestAttritionInfo 1 5 
+  [ (SubjectHasNoIndex, 0)
+  , (ExcludedBy (1, "feat2"), 3)
+  , (Included, 2)]
+  
 testAttr1p2 :: AttritionInfo
-testAttr1p2 = MkAttritionInfo 2 7 $ fromList
-  [ MkAttritionLevel SubjectHasNoIndex         0
-  , MkAttritionLevel (ExcludedBy (1, "feat2")) 4
-  , MkAttritionLevel Included                  3
-  ]
+testAttr1p2 = makeTestAttritionInfo 2 7 
+  [ (SubjectHasNoIndex, 0)
+  , (ExcludedBy (1, "feat2"), 4)
+  , (Included, 3)]
 
 tests :: TestTree
 tests = testGroup